refactor: remover debug e atualizar importações do hook useAgentsData para novo caminho
